Navbar code refactoring done
