+++ packages/frontend/src/components/Navbar/cartDrawerImports.ts
@@ -1,0 +1,6 @@
+export { useTheme } from "@mui/material/styles";
+export { default as CloseIcon } from "@mui/icons-material/Close";
+export { default as AddIcon } from "@mui/icons-material/Add";
+export { default as RemoveIcon } from "@mui/icons-material/Remove";
+export { useCart } from "../../Context/useCart";
+export { DrawerContent, ItemContainer } from "./CartDrawerStyles";
